Re organize zod schemas
